+++ AdventOfCode/TowerRow.cs
@@ -5,12 +5,12 @@ namespace AdventOfCode
 {
     public class TowerRow
     {
-        public TowerRow(int width, int height)
+        public TowerRow(int width, long height)
         {
             FilledStates = Enumerable.Repeat(false, width).ToList();
             Height = height;
         }
-        public int Height { get; private set; }
+        public long Height { get; private set; }
         public List<bool> FilledStates { get; private set; }
 
         public override string ToString()
